Create MessageContent class that manipulates messages
