+++ client/message/view.jsx
@@ -1,25 +1,8 @@
 const React = require('react/addons'),
       Avatar = require('../avatar.jsx'),
-      escape = require('escape-html'),
-      emoticons = require('emoticons'),
-      autolinks = require('autolinks');
+      MessageContent = require('./content.jsx');
 
 const Message = React.createClass({
-    _formatMessage(message) {
-        var html = escape(message);
-
-        html = emoticons.replace(html);
-
-        return {
-            __html: autolinks(html, (title, url) => {
-                return `<a href="${url}"
-                           target="_blank"
-                           rel="nofollow">
-                            ${title}
-                        </a>`;
-            })
-        };
-    },
     render() {
         var className, text = this.props.text;
 
@@ -42,9 +25,8 @@ const Message = React.createClass({
             <li className={className}>
                 <Avatar username={this.props.username} />
                 <strong>{this.props.username}</strong>
-
-                <div dangerouslySetInnerHTML={this._formatMessage(text)}>
-                </div>
+                <MessageContent messageType={this.props.messageType}
+                                message_content={message_content} />
             </li>
         );
     }
